+++ includes/vec3.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vec3.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ncarvalh <ncarvalh@student.42.fr>          +#+  +:+       +#+        */
+/*   By: crypto <crypto@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/29 14:44:02 by ncarvalh          #+#    #+#             */
-/*   Updated: 2023/07/07 23:42:31 by ncarvalh         ###   ########.fr       */
+/*   Updated: 2023/08/03 13:06:29 by crypto           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,20 +15,20 @@
 
 typedef struct s_vec3
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_vec3;
 
-t_vec3	vec3_new(float x, float y, float z);
+t_vec3	vec3_new(double x, double y, double z);
 
 t_vec3	vec3_add(t_vec3 v1, t_vec3 v2);
 
 t_vec3	vec3_sub(t_vec3 v1, t_vec3 v2);
 
-float	vec3_module(t_vec3 vec);
+double	vec3_module(t_vec3 vec);
 
-float	vec3_dot(t_vec3 v1, t_vec3 v2);
+double	vec3_dot(t_vec3 v1, t_vec3 v2);
 
 t_vec3	vec3_cross(t_vec3 v1, t_vec3 v2);
 
@@ -36,6 +36,6 @@ t_vec3	vec3_normalize(t_vec3 vec);
 
 void	vec3_print(t_vec3 vec);
 
-t_vec3	vec3_scale(t_vec3 v1, float scalar);
+t_vec3	vec3_scale(t_vec3 v1, double scalar);
 
 #endif
