+++ sources/vec3/vec3_dot.c
@@ -12,9 +12,9 @@
 
 #include "miniRT.h"
 
-float	vec3_dot(t_vec3 v1, t_vec3 v2)
+double	vec3_dot(t_vec3 v1, t_vec3 v2)
 {
-	float	dot;
+	double	dot;
 
 	dot = v1.x * v2.x + v1.y * v2.y + v1.z * v2.z;
 	return (dot);
